split monolithic app into separate modules
